Rename file from gitgub-logo.tsx to github-logo.tsx and update imports in SignIn and SignUp components.
