+++ src/auth/components/sign-in.tsx
@@ -1,6 +1,6 @@
 import { useAuthActions } from "@convex-dev/auth/react";
 import { Button } from "@/core/components/button";
-import { GitHubLogo } from "./gitgub-logo";
+import { GitHubLogo } from "./github-logo";
 import { GoogleLogo } from "./google-logo";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/core/components/card";
 import { Input } from "@/core/components/input";
